Centralize code about o flag into commonflags pkg
