+++ pkg/odo/cli/list/component/list.go
@@ -10,8 +10,8 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/redhat-developer/odo/pkg/api"
-	"github.com/redhat-developer/odo/pkg/machineoutput"
 	"github.com/redhat-developer/odo/pkg/odo/cli/ui"
+	"github.com/redhat-developer/odo/pkg/odo/commonflags"
 
 	dfutil "github.com/devfile/library/pkg/util"
 
@@ -148,7 +148,7 @@ func NewCmdComponentList(name, fullName string) *cobra.Command {
 	listCmd.Flags().StringVar(&o.namespaceFlag, "namespace", "", "Namespace for odo to scan for components")
 
 	completion.RegisterCommandFlagHandler(listCmd, "path", completion.FileCompletionHandler)
-	machineoutput.UsedByCommand(listCmd)
+	commonflags.UseOutputFlag(listCmd)
 
 	return listCmd
 }
