+++ pkg/odo/genericclioptions/runnable.go
@@ -3,8 +3,6 @@ package genericclioptions
 import (
 	"context"
 	"encoding/json"
-	"errors"
-	"flag"
 	"fmt"
 	"os"
 	"os/exec"
@@ -18,6 +16,7 @@ import (
 	"github.com/redhat-developer/odo/pkg/machineoutput"
 
 	"github.com/redhat-developer/odo/pkg/odo/cmdline"
+	"github.com/redhat-developer/odo/pkg/odo/commonflags"
 	"github.com/redhat-developer/odo/pkg/odo/genericclioptions/clientset"
 	commonutil "github.com/redhat-developer/odo/pkg/util"
 
@@ -34,7 +33,6 @@ import (
 	scontext "github.com/redhat-developer/odo/pkg/segment/context"
 
 	"github.com/spf13/cobra"
-	"github.com/spf13/pflag"
 
 	"github.com/redhat-developer/odo/pkg/log"
 	"github.com/redhat-developer/odo/pkg/odo/util"
@@ -113,7 +111,7 @@ func GenericRun(o Runnable, cmd *cobra.Command, args []string) {
 
 	// CheckMachineReadableOutput
 	// fixes / checks all related machine readable output functions
-	util.LogErrorAndExit(CheckMachineReadableOutputCommand(cmd), "")
+	util.LogErrorAndExit(commonflags.CheckMachineReadableOutputCommand(cmd), "")
 
 	deps, err := clientset.Fetch(cmd)
 	if err != nil {
@@ -187,51 +185,6 @@ func startTelemetry(cmd *cobra.Command, err error, startTime time.Time) {
 			return
 		}
 	}
-}
-
-// CheckMachineReadableOutputCommand performs machine-readable output functions required to
-// have it work correctly
-func CheckMachineReadableOutputCommand(cmd *cobra.Command) error {
-
-	// Get the needed values
-	outputFlag := pflag.Lookup("o")
-	hasFlagChanged := outputFlag != nil && outputFlag.Changed
-	machineOutput := cmd.Annotations["machineoutput"]
-
-	// Check the valid output
-	if hasFlagChanged && outputFlag.Value.String() != "json" {
-		//revive:disable:error-strings This is a top-level error message displayed as is to the end user
-		return errors.New("Please input a valid output format for -o, available format: json")
-		//revive:enable:error-strings
-	}
-
-	// Check that if -o json has been passed, that the command actually USES json.. if not, error out.
-	if hasFlagChanged && outputFlag.Value.String() == "json" && machineOutput == "" {
-
-		// By default we "disable" logging, so activate it so that the below error can be shown.
-		_ = flag.Set("o", "")
-
-		// Return the error
-		//revive:disable:error-strings This is a top-level error message displayed as is to the end user
-		return errors.New("Machine readable output is not yet implemented for this command")
-		//revive:enable:error-strings
-	}
-
-	// Before running anything, we will make sure that no verbose output is made
-	// This is a HACK to manually override `-v 4` to `-v 0` (in which we have no klog.V(0) in our code...
-	// in order to have NO verbose output when combining both `-o json` and `-v 4` so json output
-	// is not malformed / mixed in with normal logging
-	if log.IsJSON() {
-		_ = flag.Set("v", "0")
-	} else {
-		// Override the logging level by the value (if set) by the ODO_LOG_LEVEL env
-		// The "-v" flag set on command line will take precedence over ODO_LOG_LEVEL env
-		v := flag.CommandLine.Lookup("v").Value.String()
-		if level, ok := os.LookupEnv("ODO_LOG_LEVEL"); ok && v == "0" {
-			_ = flag.CommandLine.Set("v", level)
-		}
-	}
-	return nil
 }
 
 // NoArgsAndSilenceJSON returns the NoArgs value, and silence output when JSON output is activated
