+++ pkg/odo/util/flag.go
@@ -7,8 +7,6 @@ import (
 const (
 	// ProjectFlagName is the name of the flag allowing a user to specify which project to operate on
 	ProjectFlagName = "project"
-	// OutputFlagName is the name of the flag allowing user to specify output format
-	OutputFlagName = "o"
 	// ContextFlagName is the name of the flag allowing a user to specify the location of the component settings
 	ContextFlagName = "context"
 	// ComponentNameFlagName is the name of the flag allowing a user to specify which component to operate on
